# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type KeywordType = typeof KeywordType[keyof typeof KeywordType];


export const KeywordType = {
  brand: 'brand',
  competitor: 'competitor',
  keyword: 'keyword',
} as const;

export interface Keyword {
  id: number;
  text: string;
  type: KeywordType;
  createdAt: string;
}

export type KeywordInputType = typeof KeywordInputType[keyof typeof KeywordInputType];


export const KeywordInputType = {
  brand: 'brand',
  competitor: 'competitor',
  keyword: 'keyword',
} as const;

export interface KeywordInput {
  /** @minLength 1 */
  text: string;
  type: KeywordInputType;
}

export type MentionSentiment = typeof MentionSentiment[keyof typeof MentionSentiment];


export const MentionSentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export interface Mention {
  id: number;
  keywordId: number;
  platform: string;
  title: string;
  /** @nullable */
  body?: string | null;
  url: string;
  author: string;
  /** @nullable */
  subreddit?: string | null;
  sentiment: MentionSentiment;
  /** @nullable */
  sentimentScore?: number | null;
  upvotes: number;
  isUrgent: boolean;
  isComplaint: boolean;
  mentionedAt: string;
  createdAt: string;
}

export interface MentionList {
  mentions: Mention[];
  total: number;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  negative_mention: 'negative_mention',
  high_engagement: 'high_engagement',
  competitor_spike: 'competitor_spike',
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  /** @nullable */
  mentionId?: number | null;
  /** @nullable */
  keywordId?: number | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface Summary {
  id: number;
  date: string;
  content: string;
  topMentions: string;
  mainComplaints: string;
  trendingDiscussions: string;
  sentimentOverview: string;
  createdAt: string;
}

export interface ScanResult {
  scanned: number;
  newMentions: number;
  newAlerts: number;
}

export interface DashboardStats {
  totalMentions: number;
  positiveMentions: number;
  neutralMentions: number;
  negativeMentions: number;
  urgentMentions: number;
  complaintMentions: number;
  unreadAlerts: number;
  trackedKeywords: number;
}

export interface SentimentTrendPoint {
  date: string;
  positive: number;
  neutral: number;
  negative: number;
}

export type ListMentionsParams = {
keywordId?: number;
sentiment?: ListMentionsSentiment;
platform?: string;
limit?: number;
offset?: number;
};

export type ListMentionsSentiment = typeof ListMentionsSentiment[keyof typeof ListMentionsSentiment];


export const ListMentionsSentiment = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export type ListAlertsParams = {
unreadOnly?: boolean;
limit?: number;
};

export type ListSummariesParams = {
limit?: number;
};