# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /keywords:
    get:
      operationId: listKeywords
      summary: List all tracked keywords
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      summary: Create a tracked keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /keywords/{id}:
    delete:
      operationId: deleteKeyword
      summary: Delete a tracked keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Keyword:
      type: object
      required: [id, text, type, createdAt]
      properties:
        id:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [brand, competitor, keyword]
        createdAt:
          type: string
          format: date-time
    KeywordInput:
      type: object
      required: [text, type]
      properties:
        text:
          type: string
          minLength: 1
        type:
          type: string
          enum: [brand, competitor, keyword]
...