# lib/api-zod/src/generated/types/keyword.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { KeywordType } from './keywordType';

export interface Keyword {
  id: number;
  text: string;
  type: KeywordType;
  createdAt: Date;
}