# lib/api-zod/src/generated/types/keywordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { KeywordInputType } from './keywordInputType';

export interface KeywordInput {
  /** @minLength 1 */
  text: string;
  type: KeywordInputType;
}